de-dup constants and split huge Bank file into a few separate
