+++ programs/marginfi/fuzz/src/account_state.rs
@@ -13,7 +13,7 @@ use anchor_spl::token_2022::spl_token_2022::{
     state::Mint,
 };
 use bumpalo::Bump;
-use marginfi::{constants::FEE_STATE_SEED, state::bank::BankVaultType};
+use marginfi::{state::bank::BankVaultType};
 use pyth_solana_receiver_sdk::price_update::{PriceUpdateV2, VerificationLevel};
 use safe_transmute::transmute_to_bytes_mut;
 use solana_program::{
@@ -21,6 +21,7 @@ use solana_program::{
 };
 use solana_sdk::{signature::Keypair, signer::Signer};
 use std::mem::size_of;
+use marginfi_type_crate::constants::FEE_STATE_SEED;
 
 pub struct AccountsState {
     pub bump: Bump,
