+++ programs/marginfi/src/constants.rs
@@ -1,22 +1,9 @@
 use anchor_lang::prelude::*;
-use anchor_lang::solana_program::pubkey;
-use fixed::types::I80F48;
-use fixed_macro::types::I80F48;
 use pyth_solana_receiver_sdk::price_update::VerificationLevel;
 
-pub const LIQUIDITY_VAULT_AUTHORITY_SEED: &str = "liquidity_vault_auth";
-pub const INSURANCE_VAULT_AUTHORITY_SEED: &str = "insurance_vault_auth";
-pub const FEE_VAULT_AUTHORITY_SEED: &str = "fee_vault_auth";
-
-pub const LIQUIDITY_VAULT_SEED: &str = "liquidity_vault";
-pub const INSURANCE_VAULT_SEED: &str = "insurance_vault";
-pub const FEE_VAULT_SEED: &str = "fee_vault";
-
-pub const FEE_STATE_SEED: &str = "feestate";
-pub const STAKED_SETTINGS_SEED: &str = "staked_settings";
-
-pub const EMISSIONS_AUTH_SEED: &str = "emissions_auth_seed";
-pub const EMISSIONS_TOKEN_ACCOUNT_SEED: &str = "emissions_token_account_seed";
+// This file should only contain the constants which couldn't be moved to type-crate:
+// 1. the constants used for testing/internal purposes
+// 2. or the ones dependant on some 3rd party crates which are not part of type-crate dependency tree
 
 /// Used for the health cache to track which version of the program generated it.
 /// * 0 = invalid
@@ -57,137 +44,10 @@ cfg_if::cfg_if! {
 
 pub const NATIVE_STAKE_ID: Pubkey = pubkey!("Stake11111111111111111111111111111111111111");
 
-/// TODO: Make these variable per bank
-pub const LIQUIDATION_LIQUIDATOR_FEE: I80F48 = I80F48!(0.025);
-pub const LIQUIDATION_INSURANCE_FEE: I80F48 = I80F48!(0.025);
-
 /// The default fee, in native SOL in native decimals (i.e. lamports) used in testing
 pub const INIT_BANK_ORIGINATION_FEE_DEFAULT: u32 = 10000;
 
-pub const SECONDS_PER_YEAR: I80F48 = I80F48!(31_536_000);
-
-/// Due to real-world constraints, oracles using an age less than this value are typically too
-/// unreliable, and we want to restrict pools from picking an oracle that is effectively unusable
-pub const ORACLE_MIN_AGE: u16 = 30;
-pub const MAX_PYTH_ORACLE_AGE: u64 = 60;
-
-/// Range that contains 95% price data distribution
-///
-/// https://docs.pyth.network/price-feeds/best-practices#confidence-intervals
-pub const CONF_INTERVAL_MULTIPLE: I80F48 = I80F48!(2.12);
-/// Range that contains 95% price data distribution in a normal distribution
-pub const STD_DEV_MULTIPLE: I80F48 = I80F48!(1.96);
-/// Maximum confidence interval allowed
-pub const MAX_CONF_INTERVAL: I80F48 = I80F48!(0.05);
-pub const U32_MAX: I80F48 = I80F48!(4_294_967_295);
-pub const U32_MAX_DIV_10: I80F48 = I80F48!(429_496_730);
-
-pub const USDC_EXPONENT: i32 = 6;
-
-pub const MAX_ORACLE_KEYS: usize = 5;
-
-/// Any balance below 1 SPL token amount is treated as none,
-/// this is to account for any artifacts resulting from binary fraction arithemtic.
-pub const EMPTY_BALANCE_THRESHOLD: I80F48 = I80F48!(1);
-
-/// Any account with assets below this threshold is considered bankrupt.
-/// The account also needs to have more liabilities than assets.
-///
-/// This is USD denominated, so 0.001 = $0.1
-pub const BANKRUPT_THRESHOLD: I80F48 = I80F48!(0.1);
-
-/// Comparios threshold used to account for arithmetic artifacts on balances
-pub const ZERO_AMOUNT_THRESHOLD: I80F48 = I80F48!(0.0001);
-
-pub const EMISSIONS_FLAG_BORROW_ACTIVE: u64 = 1 << 0;
-pub const EMISSIONS_FLAG_LENDING_ACTIVE: u64 = 1 << 1;
-pub const PERMISSIONLESS_BAD_DEBT_SETTLEMENT_FLAG: u64 = 1 << 2;
-pub const FREEZE_SETTINGS: u64 = 1 << 3;
-pub const CLOSE_ENABLED_FLAG: u64 = 1 << 4;
-
-/// True if bank created in 0.1.4 or later, or if migrated to the new oracle setup from a prior
-/// version. False otherwise.
-pub const PYTH_PUSH_MIGRATED: u8 = 1 << 0;
-
-/// Cutoff timestamp for balance last_update used in accounting collected emissions.
-/// Any balance updates before this timestamp are ignored, and current_timestamp is used instead.
-pub const MIN_EMISSIONS_START_TIME: u64 = 1681989983;
-
-pub const MAX_EXP_10_I80F48: usize = 24;
-pub const EXP_10_I80F48: [I80F48; MAX_EXP_10_I80F48] = [
-    I80F48!(1),                        // 10^0
-    I80F48!(10),                       // 10^1
-    I80F48!(100),                      // 10^2
-    I80F48!(1000),                     // 10^3
-    I80F48!(10000),                    // 10^4
-    I80F48!(100000),                   // 10^5
-    I80F48!(1000000),                  // 10^6
-    I80F48!(10000000),                 // 10^7
-    I80F48!(100000000),                // 10^8
-    I80F48!(1000000000),               // 10^9
-    I80F48!(10000000000),              // 10^10
-    I80F48!(100000000000),             // 10^11
-    I80F48!(1000000000000),            // 10^12
-    I80F48!(10000000000000),           // 10^13
-    I80F48!(100000000000000),          // 10^14
-    I80F48!(1000000000000000),         // 10^15
-    I80F48!(10000000000000000),        // 10^16
-    I80F48!(100000000000000000),       // 10^17
-    I80F48!(1000000000000000000),      // 10^18
-    I80F48!(10000000000000000000),     // 10^19
-    I80F48!(100000000000000000000),    // 10^20
-    I80F48!(1000000000000000000000),   // 10^21
-    I80F48!(10000000000000000000000),  // 10^22
-    I80F48!(100000000000000000000000), // 10^23
-];
-
-pub const MAX_EXP_10: usize = 21;
-pub const EXP_10: [i128; MAX_EXP_10] = [
-    1,                     // 10^0
-    10,                    // 10^1
-    100,                   // 10^2
-    1000,                  // 10^3
-    10000,                 // 10^4
-    100000,                // 10^5
-    1000000,               // 10^6
-    10000000,              // 10^7
-    100000000,             // 10^8
-    1000000000,            // 10^9
-    10000000000,           // 10^10
-    100000000000,          // 10^11
-    1000000000000,         // 10^12
-    10000000000000,        // 10^13
-    100000000000000,       // 10^14
-    1000000000000000,      // 10^15
-    10000000000000000,     // 10^16
-    100000000000000000,    // 10^17
-    1000000000000000000,   // 10^18
-    10000000000000000000,  // 10^19
-    100000000000000000000, // 10^20
-];
-
-/// Value where total_asset_value_init_limit is considered inactive
-pub const TOTAL_ASSET_VALUE_INIT_LIMIT_INACTIVE: u64 = 0;
-
-/// For testing, this is a typical program fee.
-pub const PROTOCOL_FEE_RATE_DEFAULT: I80F48 = I80F48!(0.025);
-/// For testing, this is a typical program fee.
-pub const PROTOCOL_FEE_FIXED_DEFAULT: I80F48 = I80F48!(0.01);
-
 pub const MIN_PYTH_PUSH_VERIFICATION_LEVEL: VerificationLevel = VerificationLevel::Full;
-/// Pyth Pull Oracles sponsored by Pyth use this shard ID.
-pub const PYTH_SPONSORED_SHARD_ID: u16 = 0;
-/// Pyth Pull Oracles sponsored by Marginfi use this shard ID.
-pub const MARGINFI_SPONSORED_SHARD_ID: u16 = 3301;
-
-/// A regular asset that can be comingled with any other regular asset or with `ASSET_TAG_SOL`
-pub const ASSET_TAG_DEFAULT: u8 = 0;
-/// Accounts with a SOL position can comingle with **either** `ASSET_TAG_DEFAULT` or
-/// `ASSET_TAG_STAKED` positions, but not both
-pub const ASSET_TAG_SOL: u8 = 1;
-/// Staked SOL assets. Accounts with a STAKED position can only deposit other STAKED assets or SOL
-/// (`ASSET_TAG_SOL`) and can only borrow SOL (`ASSET_TAG_SOL`)
-pub const ASSET_TAG_STAKED: u8 = 2;
 
 // TODO move this to the global fee wallet eventually
 /// A nominal fee paid to the global wallet when intiating an account transfer. Primarily intended
