+++ programs/marginfi/src/instructions/marginfi_group/add_pool_permissionless.rs
@@ -2,25 +2,25 @@
 // stake pool to a group so users can borrow SOL against it
 use crate::{
     check,
-    constants::{
-        ASSET_TAG_STAKED, FEE_VAULT_AUTHORITY_SEED, FEE_VAULT_SEED, INSURANCE_VAULT_AUTHORITY_SEED,
-        INSURANCE_VAULT_SEED, LIQUIDITY_VAULT_AUTHORITY_SEED, LIQUIDITY_VAULT_SEED,
-        PYTH_PUSH_MIGRATED, SPL_SINGLE_POOL_ID,
-    },
+    constants::SPL_SINGLE_POOL_ID,
     events::{GroupEventHeader, LendingPoolBankCreateEvent},
     log_pool_info,
-    state::{
-        bank::{BankConfigImpl, BankImpl},
-        marginfi_group::MarginfiGroupImpl,
-    },
+    state::{bank::BankImpl, bank_config::BankConfigImpl, marginfi_group::MarginfiGroupImpl},
     MarginfiError, MarginfiResult,
 };
 use anchor_lang::prelude::*;
 use anchor_spl::token_interface::*;
 use fixed_macro::types::I80F48;
-use marginfi_type_crate::types::{
-    Bank, BankConfigCompact, BankOperationalState, InterestRateConfig, MarginfiGroup, OracleSetup,
-    StakedSettings,
+use marginfi_type_crate::{
+    constants::{
+        ASSET_TAG_STAKED, FEE_VAULT_AUTHORITY_SEED, FEE_VAULT_SEED, INSURANCE_VAULT_AUTHORITY_SEED,
+        INSURANCE_VAULT_SEED, LIQUIDITY_VAULT_AUTHORITY_SEED, LIQUIDITY_VAULT_SEED,
+        PYTH_PUSH_MIGRATED,
+    },
+    types::{
+        Bank, BankConfigCompact, BankOperationalState, InterestRateConfig, MarginfiGroup,
+        OracleSetup, StakedSettings,
+    },
 };
 
 pub fn lending_pool_add_bank_permissionless(
