+++ programs/marginfi/src/instructions/marginfi_group/close_bank.rs
@@ -1,13 +1,13 @@
 use crate::check;
 use crate::state::bank::BankImpl;
 use crate::utils::NumTraitsWithTolerance;
-use crate::{
-    constants::{CLOSE_ENABLED_FLAG, ZERO_AMOUNT_THRESHOLD},
-    MarginfiError, MarginfiResult,
-};
+use crate::{MarginfiError, MarginfiResult};
 use anchor_lang::prelude::*;
 use fixed::types::I80F48;
-use marginfi_type_crate::types::{Bank, MarginfiGroup};
+use marginfi_type_crate::{
+    constants::{CLOSE_ENABLED_FLAG, ZERO_AMOUNT_THRESHOLD},
+    types::{Bank, MarginfiGroup},
+};
 
 pub fn lending_pool_close_bank(ctx: Context<LendingPoolCloseBank>) -> MarginfiResult {
     let mut group = ctx.accounts.group.load_mut()?;
