+++ programs/marginfi/src/instructions/marginfi_group/collect_bank_fees.rs
@@ -1,19 +1,20 @@
-use crate::constants::{FEE_STATE_SEED, FEE_VAULT_AUTHORITY_SEED, INSURANCE_VAULT_AUTHORITY_SEED};
 use crate::events::{GroupEventHeader, LendingPoolBankCollectFeesEvent};
 use crate::state::bank::{BankImpl, BankVaultType};
-use crate::{
-    bank_signer,
-    constants::{
-        FEE_VAULT_SEED, INSURANCE_VAULT_SEED, LIQUIDITY_VAULT_AUTHORITY_SEED, LIQUIDITY_VAULT_SEED,
-    },
-    math_error, MarginfiResult,
-};
+use crate::{bank_signer, math_error, MarginfiResult};
 use crate::{check, utils, MarginfiError};
 use anchor_lang::prelude::*;
 use anchor_spl::associated_token::get_associated_token_address_with_program_id;
 use anchor_spl::token_interface::{TokenAccount, TokenInterface};
 use fixed::types::I80F48;
-use marginfi_type_crate::types::{Bank, FeeState, MarginfiGroup};
+use marginfi_type_crate::constants::{
+    FEE_STATE_SEED, FEE_VAULT_AUTHORITY_SEED, INSURANCE_VAULT_AUTHORITY_SEED,
+};
+use marginfi_type_crate::{
+    constants::{
+        FEE_VAULT_SEED, INSURANCE_VAULT_SEED, LIQUIDITY_VAULT_AUTHORITY_SEED, LIQUIDITY_VAULT_SEED,
+    },
+    types::{Bank, FeeState, MarginfiGroup},
+};
 use std::cmp::min;
 
 pub fn lending_pool_collect_bank_fees<'info>(
