+++ programs/marginfi/src/instructions/marginfi_group/configure_bank_lite.rs
@@ -1,9 +1,12 @@
 use crate::set_if_some;
-use crate::state::bank::InterestRateConfigImpl;
+use crate::state::bank::BankImpl;
+use crate::state::interest_rate::InterestRateConfigImpl;
 use crate::MarginfiResult;
-use crate::{constants::FREEZE_SETTINGS, state::bank::BankImpl};
 use anchor_lang::prelude::*;
-use marginfi_type_crate::types::{Bank, InterestRateConfigOpt, MarginfiGroup};
+use marginfi_type_crate::{
+    constants::FREEZE_SETTINGS,
+    types::{Bank, InterestRateConfigOpt, MarginfiGroup},
+};
 
 pub fn lending_pool_configure_bank_interest_only(
     ctx: Context<LendingPoolConfigureBankInterestOnly>,
