+++ programs/marginfi/src/instructions/marginfi_group/init_staked_settings.rs
@@ -1,8 +1,11 @@
 // Used by the group admin to enable staked collateral banks and configure their default features
+use crate::state::staked_settings::StakedSettingsImpl;
 use crate::utils::wrapped_i80f48_to_f64;
-use crate::{constants::STAKED_SETTINGS_SEED, state::staked_settings::StakedSettingsImpl};
 use anchor_lang::prelude::*;
-use marginfi_type_crate::types::{MarginfiGroup, RiskTier, StakedSettings, WrappedI80F48};
+use marginfi_type_crate::{
+    constants::STAKED_SETTINGS_SEED,
+    types::{MarginfiGroup, RiskTier, StakedSettings, WrappedI80F48},
+};
 
 pub fn initialize_staked_settings(
     ctx: Context<InitStakedSettings>,
