+++ programs/marginfi/tests/misc/collateral_value_cap.rs
@@ -1,6 +1,6 @@
 use fixtures::{assert_custom_error, prelude::*};
-use marginfi::{constants::TOTAL_ASSET_VALUE_INIT_LIMIT_INACTIVE, prelude::MarginfiError};
-use marginfi_type_crate::types::BankConfigOpt;
+use marginfi::prelude::MarginfiError;
+use marginfi_type_crate::{constants::TOTAL_ASSET_VALUE_INIT_LIMIT_INACTIVE, types::BankConfigOpt};
 use pretty_assertions::assert_eq;
 use solana_program_test::*;
