+++ programs/marginfi/tests/misc/regression.rs
@@ -6,9 +6,11 @@ use anyhow::bail;
 use base64::{prelude::BASE64_STANDARD, Engine};
 use bytemuck::Zeroable;
 use fixed::types::I80F48;
-use marginfi::constants::ASSET_TAG_DEFAULT;
-use marginfi_type_crate::types::{
-    Bank, BankCache, BankOperationalState, HealthCache, MarginfiAccount, OracleSetup, RiskTier,
+use marginfi_type_crate::{
+    constants::ASSET_TAG_DEFAULT,
+    types::{
+        Bank, BankCache, BankOperationalState, HealthCache, MarginfiAccount, OracleSetup, RiskTier,
+    },
 };
 use solana_account_decoder::UiAccountData;
 use solana_cli_output::CliAccount;
